requirements/base.txt
Django==3.2.7
django-model-utils==4.1.1